Add default to `id` field of shared CLI via const literal environment variable
yes it does work... The shared CLI, `morivar::cli::ClientArguments`, has a field `id`, which should be the binary name unless specified.
This field was previously set with `unwrap_or(_else)` on `Option<String>`. But that doesn't display nicely in clap.
Now, it is a const literal `&'static str` of the shared CLI, which can be set by calling `{ env!("CARGO_BIN_NAME") }` at the usage site.
Note, this requires turning on incomplete features of the compiler and ignoring the warnings about it.